src/requirements.txt
gradio>=4.0.0
gradio>=4.0.0[mcp] 
gradio>=4.0.0_client>=0.8.0
Pillow>=9.0.0
requests>=2.25.0
transformers>=4.35.0